[sipearl][ci] Build OpenMP runtime using LLVM_ENABLE_RUNTIMES
Fix the build of OpenMP: librt issue with -rpath-link
librt.so has a dependency on libpthread but aarch64-linux-gnu-ld is
not able to search for dynamic library dependencies using -L options.
Have to pass -rpath-link for right dynamic librairies dependencies
search path. But cmake check for support of shm_open using librt
in phase of cross compilation... For support of librt as it is
complex to give correct cmake option from top level to cross
compilation phase.