[flang][PRIF] Add allocation for coarrays with PRIF
First version of static coarray allocation and with the ALLOCATE statement.
A temporary version of the mechanism for saving the COARRAY_HANDLE has
been made. In the futur the idea is to remove the use of the std::map implemented at
runtime level and attach the HANDLE directly on the coarray.