[flang][PRIF] Adding support of -fcoarray option
Adding support of -fcoarray (and -fcoarray=lib as an alias of -fcoarray).
If this option is not passed, then coarray features are disabled and error is
returned if a least one feature is present at compile time.

Added lowering for `prif_init` and `prif_[error_]stop` to call them in
ProgramStart and ProgramEnd.